fix false disconnect due to a signal sent to the reading process
- In the current implementation, when a signal is sent to the reading
process,
read is cancelled by calling usb_ep_dequeue, which lead into calling
acc_complete_out with ECONNRESET, but the current logic treats it as
disconnection, which makes the device inaccessible until cable is
actually
disconnected.
- The fix calls disconnect only when ESHUTDOWN error is passed.
- If data has already arrived while trying cancelling, the data is
marked
as available, and it will be read out on the next read. This is
necessary
as USB bulk is assumed to guarantee no data loss.

Change-Id: I4e0da4c5ec43e91b692ed8c0053e43b6f047e52f
Signed-off-by: keunyoung <keunyoung@google.com>