touch_synaptics: i2c_read: auto retry 10 times before erroring out
The i2c devices we are connecting to may not be ready yet, or just busy.
Retry with a 10ms pause in order to still obtain touch data.
This improves overall touch data robustness, but the main thing it does:
WE CAN FINALLY GET TOUCH DATA IN SUSPEND ALL THE TIME.

Signed-off-by:
Dennis Rassmann <showp1984@gmail.com>